requirements.txt
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pypdfium2==5.6.0
pytest==9.0.2
requests==2.32.5
